consumer/requirements.txt
kafka-python    # to consume messages from Kafka
pymongo         # to connect to MongoDB and insert data
